Rename/move utility to reflect new scope
The renderer can now be used for both Application and Assessment
documents, so it makes sense to move it out of the application-specific
directory, and to rename it to somehting more generic.
