+++ server/views/applications/partials/_readonly-application.njk
@@ -3,23 +3,21 @@
 
 {% macro applicationReadonlyView(application) %}
 
-  {% if application.status === 'withdrawn' %}
-    {% set sections = SummaryListUtils.summaryListSections(application, false) %}
-  {%else%}
-    {% set sections = getSummaryCardsForApplication(application) %}
-  {%endif%}
+    {% if application.status === 'withdrawn' %}
+        {% set sections = SummaryListUtils.summaryListSections(application, false) %}
+    {% else %}
+        {% set sections = getDocumentSections(application) %}
+    {% endif %}
 
     {{ personDetails(application.person, statusOnSubmission = application.personStatusOnSubmission) }}
 
     {% if application.type !== 'Offline' %}
-      {% for section in sections %}
-        <h2 class="govuk-heading-l">{{ section.title }}</h2>
-        {% for task in section.tasks %}
-          {{
-              govukSummaryList(task)
-            }}
+        {% for section in sections %}
+            <h2 class="govuk-heading-l">{{ section.title }}</h2>
+            {% for task in section.tasks %}
+                {{ govukSummaryList(task) }}
+            {% endfor %}
         {% endfor %}
-      {% endfor %}
     {% endif %}
 
-  {% endmacro %}
+{% endmacro %}
