+++ server/views/assessments/show.njk
@@ -27,7 +27,7 @@
 
             {{ personDetails(assessment.application.person, statusOnSubmission = assessment.application.personStatusOnSubmission) }}
 
-            {% for section in getSummaryCardsForApplication(assessment) %}
+            {% for section in getDocumentSections(assessment) %}
                 <h2 class="govuk-heading-l">{{ section.title }}</h2>
                 {% for task in section.tasks %}
                     {{ govukSummaryList(task) }}
