+++ server/views/manage/premises/placements/show.njk
@@ -73,7 +73,7 @@
 
     {% elseif activeTab==='assessment' %}
 
-        {% for section in getSummaryCardsForApplication(assessment) %}
+        {% for section in getDocumentSections(assessment) %}
             <h2 class="govuk-heading-l">{{ section.title }}</h2>
             {% for task in section.tasks %}
                 {{ govukSummaryList(task) }}
